pyproject.toml
[tool.poetry]
name = "fastapi-jwt-auth"
version = "0.1.0"
description = ""
authors = ["sanfit <sanfit@vk.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
fastapi = "^0.111.0"
alembic = "^1.13.1"
pydantic-settings = "^2.3.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
pytest = "^8.2.1"
black = "^24.4.2"
pyright = "^1.1.365"
pyjwt = "^2.8.0"
bcrypt = "^4.1.3"
cryptography = "^42.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
